streamlined title update process and aligned it with design
