+++ app/src/main/res/layout/activity_main.xml
@@ -16,12 +16,12 @@
         app:layout_constraintTop_toTopOf="parent">
 
         <TextView
+            style="?textAppearanceHeadline5"
             android:id="@+id/headerTitle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_margin="@dimen/dimen_5_dp"
             android:layout_marginStart="8dp"
-            android:padding="@dimen/dimen_5_dp"
             android:text="@string/app_name"
             android:textColor="@color/colorBlack"
             android:textSize="20sp" />
