+++ commonUI/src/main/java/org/navgurukul/commonui/platform/ToolbarConfigurable.kt
@@ -1,8 +1,11 @@
 package org.navgurukul.commonui.platform
 
+import androidx.annotation.AttrRes
 import androidx.appcompat.widget.Toolbar
 
 interface ToolbarConfigurable {
 
     fun configure(toolbar: Toolbar)
+
+    fun setTitle(title: String, @AttrRes colorRes: Int)
 }
