+++ commonUI/src/main/java/org/navgurukul/commonui/themes/ThemeUtils.kt
@@ -1,0 +1,99 @@
+package org.navgurukul.commonui.themes
+
+import android.content.Context
+import android.graphics.drawable.Drawable
+import android.util.TypedValue
+import android.view.Menu
+import androidx.annotation.AttrRes
+import androidx.annotation.ColorInt
+import androidx.core.content.ContextCompat
+import androidx.core.graphics.drawable.DrawableCompat
+import timber.log.Timber
+
+/**
+ * Util class for managing themes.
+ */
+object ThemeUtils {
+    private val mColorByAttr = HashMap<Int, Int>()
+
+
+    /**
+     * Translates color attributes to colors
+     *
+     * @param c              Context
+     * @param colorAttribute Color Attribute
+     * @return Requested Color
+     */
+    @ColorInt
+    fun getColor(c: Context, @AttrRes colorAttribute: Int): Int {
+        return mColorByAttr.getOrPut(colorAttribute) {
+            try {
+                val color = TypedValue()
+                c.theme.resolveAttribute(colorAttribute, color, true)
+                color.data
+            } catch (e: Exception) {
+                Timber.e(e, "Unable to get color")
+                ContextCompat.getColor(c, android.R.color.holo_red_dark)
+            }
+        }
+    }
+
+    fun getAttribute(c: Context, @AttrRes attribute: Int): TypedValue? {
+        try {
+            val typedValue = TypedValue()
+            c.theme.resolveAttribute(attribute, typedValue, true)
+            return typedValue
+        } catch (e: Exception) {
+            Timber.e(e, "Unable to get color")
+        }
+        return null
+    }
+
+    /**
+     * Update the menu icons colors
+     *
+     * @param menu  the menu
+     * @param color the color
+     */
+    fun tintMenuIcons(menu: Menu, color: Int) {
+        for (i in 0 until menu.size()) {
+            val item = menu.getItem(i)
+            val drawable = item.icon
+            if (drawable != null) {
+                val wrapped = DrawableCompat.wrap(drawable)
+                drawable.mutate()
+                DrawableCompat.setTint(wrapped, color)
+                item.icon = drawable
+            }
+        }
+    }
+
+    /**
+     * Tint the drawable with a theme attribute
+     *
+     * @param context   the context
+     * @param drawable  the drawable to tint
+     * @param attribute the theme color
+     * @return the tinted drawable
+     */
+    fun tintDrawable(context: Context, drawable: Drawable, @AttrRes attribute: Int): Drawable {
+        return tintDrawableWithColor(drawable, getColor(context, attribute))
+    }
+
+    /**
+     * Tint the drawable with a color integer
+     *
+     * @param drawable the drawable to tint
+     * @param color    the color
+     * @return the tinted drawable
+     */
+    fun tintDrawableWithColor(drawable: Drawable, @ColorInt color: Int): Drawable {
+        val tinted = DrawableCompat.wrap(drawable)
+        drawable.mutate()
+        DrawableCompat.setTint(tinted, color)
+        return tinted
+    }
+}
+
+fun Context.getThemedColor(@AttrRes attribute: Int) = ThemeUtils.getColor(this, attribute)
+
