+++ learn/build.gradle
@@ -24,6 +24,7 @@ android {
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation project(":core")
+    implementation project(path: ':commonUI')
     implementation "androidx.core:core-ktx:$core_ktx_verision"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
